Fix typo, rename improts module -> imports
